refactor: Rename /materials to /Material
