+++ public/minecraft/scenes/instancing.js
@@ -1,6 +1,6 @@
 import {Camera, Mesh} from "../../../src/index.js";
 import {BoxGeometry} from "../../../src/Geometry/index.js";
-import {Material} from "../../../src/materials/index.js";
+import {Material} from "../../../src/Material/index.js";
 import {Matrix3, PI, Vector2, Vector3} from "../../../src/math/index.js";
 import {BLOCK_SCALE, FIELD_OF_VIEW} from "../main.js";
 import {Scene} from "../Scene.js";
